# docs/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cognitive Quorum API",
    "description": "Backend for Cognitive Quorum application.",
    "version": "2.1.0"
  },
  "paths": {
...
    "/llm/llm/completion": {
      "post": {
        "tags": [
          "LLM",
          "LLM"
        ],
        "summary": "Direct Completion",
        "description": "Directly invokes the LLM using the specified strategy.\n\nSupports structured output if schema is provided.\n\nArgs:\n    request (CompletionRequest): The prompt and settings.\n    registry (RegistryDep): Registry dependency to resolve strategies.\n\nReturns:\n    dict: Result object containing the generated content.\n\nRaises:\n    HTTPException: If strategy is invalid (400) or generation fails (500).",
        "operationId": "generate_completion_llm_llm_completion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The generated text or structured object.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm/llm/batch-completion": {
      "post": {
        "tags": [
          "LLM",
          "LLM"
        ],
        "summary": "Batch Completion",
        "description": "Processes multiple completion requests in parallel.\n\nArgs:\n    batch (BatchCompletionRequest): List of requests.\n    registry (RegistryDep): Registry dependency.\n\nReturns:\n    dict: List of results (success or error) for each request.",
        "operationId": "batch_completion_llm_llm_batch_completion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of results.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/llm/llm/config": {
      "get": {
        "tags": [
          "LLM",
          "LLM"
        ],
        "summary": "Get Model Registry",
        "description": "Retrieves the active model registry, which maps abstract strategies (e.g., 'fast') to concrete models.\n\nArgs:\n    handler: Dependency.\n\nReturns:\n    dict: The registry configuration object.",
        "operationId": "get_model_config_llm_llm_config_get",
        "responses": {
          "200": {
            "description": "The current internal model mapping configuration.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "LLM",
          "LLM"
        ],
        "summary": "Update Model Registry",
        "description": "Updates the system's model registry configuration in the database.\n\nArgs:\n    update (ModelRegistryUpdate): The new configuration.\n    db_client (DatabaseDep): Database dependency.\n\nReturns:\n    dict: Status and the updated registry.\n\nRaises:\n    HTTPException: If database update fails (500).",
        "operationId": "update_model_config_llm_llm_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRegistryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Confirmation of the configuration update.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchCompletionRequest": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/CompletionRequest"
            },
            "type": "array",
            "title": "Requests",
            "description": "List of requests to process in parallel."
          }
        },
        "type": "object",
        "required": [
          "requests"
        ],
        "title": "BatchCompletionRequest",
        "description": "Payload for batch completion requests."
      },
...
      "CompletionRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "The primary prompt text."
          },
          "system_instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Instruction",
            "description": "Optional system instruction."
          },
          "model_strategy": {
            "type": "string",
            "title": "Model Strategy",
            "description": "Strategy key (fast, deep, etc) or direct model name.",
            "default": "fast"
          },
          "response_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Schema",
            "description": "Optional JSON Schema for structured output."
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "CompletionRequest",
        "description": "Payload for a single LLM completion request.\n\nAttributes:\n    prompt (str): The primary input text.\n    system_instruction (str): Optional system context.\n    model_strategy (str): Strategy key ('fast', 'deep') or direct model name.\n    response_schema (dict): JSON Schema for structured output validation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelRegistryUpdate": {
        "properties": {
          "registry": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Registry",
            "description": "The new configuration map for model strategies (e.g. {'fast': {'model_name': '...'}})."
          }
        },
        "type": "object",
        "required": [
          "registry"
        ],
        "title": "ModelRegistryUpdate",
        "description": "Payload for updating the model registry."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}